Final commit

+++ app/Http/Controllers/ListingController.php
@@ -45,6 +45,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos','public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
 //        Session::flash('message','Listing Created');
@@ -57,11 +59,19 @@ class ListingController extends Controller
     }
     public function destroy(Listing $listing)
     {
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized action');
+        }
         $listing->delete();
         return redirect('/')->with('message','Listing deleted successfully');
     }
     public function update(Request $request,Listing $listing)
     {
+
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized action');
+        }
         $formFields = $request->validate(
             [
                 'title' => 'required',
@@ -80,8 +90,13 @@ class ListingController extends Controller
 
         $listing->update($formFields);
 
+
 //        Session::flash('message','Listing Created');
 
         return back()->with('message','Listing updated successfully');
+    }
+
+    public function manage(){
+        return view('listings.manage',['listings'=>auth()->user()->listings()->get()]);
     }
 }
